brandwatch-cdk/requirements.txt
-e .
aws_cdk.aws_lambda
aws_cdk.aws_s3